Refactor evaluate to reduce file sizes.
